+++ src/face.rs
@@ -70,16 +70,14 @@ impl<'a> Face<'a> {
     /// Data will be referenced, not owned.
     pub fn from_slice(data: &'a [u8], face_index: u32) -> Option<Self> {
         let face = ttf_parser::Face::parse(data, face_index).ok()?;
-        Self::from_face(face)
+        Some(Self::from_face(face))
     }
 
     /// Creates a new [`Face`] from [`ttf_parser::Face`].
     ///
     /// Data will be referenced, not owned.
-    ///
-    /// Returns `None` when face's units per EM is `None`.
-    pub fn from_face(face: ttf_parser::Face<'a>) -> Option<Self> {
-        Some(Face {
+    pub fn from_face(face: ttf_parser::Face<'a>) -> Self {
+        Face {
             units_per_em: face.units_per_em(),
             pixels_per_em: None,
             points_per_em: None,
@@ -87,7 +85,7 @@ impl<'a> Face<'a> {
             gsub: face.tables().gsub.map(SubstitutionTable::new),
             gpos: face.tables().gpos.map(PositioningTable::new),
             ttfp_face: face,
-        })
+        }
     }
 
     // TODO: remove